换回7b

+++ langchain/run_gradio.py
@@ -9,12 +9,12 @@ os.system("git clone https://gitee.com/yzy0612/nltk_data.git  --branch gh-pages;
 
 
 #download model
-# if not os.path.exists('/home/xlab-app-center/InternLM-chat-7b-8k'):
-os.system("pip install -U openxlab")
-from openxlab.model import download
-    # download(model_repo='OpenLMLab/InternLM-chat-7b-8k',output='/home/xlab-app-center/InternLM-chat-7b-8k')
+if not os.path.exists('/home/xlab-app-center/InternLM-chat-7b-8k'):
+    os.system("pip install -U openxlab")
+    from openxlab.model import download
+    download(model_repo='OpenLMLab/InternLM-chat-7b-8k',output='/home/xlab-app-center/InternLM-chat-7b-8k')
     # YuanLLM/Yuan2-2B-hf
-download(model_repo='YuanLLM/Yuan2-2B-hf',output='/home/xlab-app-center/InternLM-chat-7b-8k')
+# download(model_repo='YuanLLM/Yuan2-2B-hf',output='/home/xlab-app-center/InternLM-chat-7b-8k')
     # os.system("pip install flash_attn")
 if not os.path.exists('/home/xlab-app-center/sentence-transformer'):
     # 设置环境变量
